+++ spec/recipes/default_spec.rb
@@ -4,11 +4,11 @@
 require_relative '../spec_helper'
 
 describe 'dnsmasq-local::default' do
-  %i[config options].each { |a| let(a) { nil } }
+  %i[config options environment].each { |a| let(a) { nil } }
   let(:platform) { { platform: 'ubuntu', version: '14.04' } }
   let(:runner) do
     ChefSpec::SoloRunner.new(platform) do |node|
-      %i[config options].each do |a|
+      %i[config options environment].each do |a|
         node.normal['dnsmasq_local'][a] = send(a) unless send(a).nil?
       end
     end
@@ -18,7 +18,9 @@ describe 'dnsmasq-local::default' do
   shared_examples_for 'any attribute set' do
     it 'creates a dnsmasq_local resource' do
       expect(chef_run).to create_dnsmasq_local('default')
-        .with(config: config || {}, options: options || {})
+        .with(config: config || {},
+              options: options || {},
+              environment: environment || {})
     end
   end
 
@@ -36,6 +38,12 @@ describe 'dnsmasq-local::default' do
 
   context 'an overridden options attribute' do
     let(:options) { { 'bind_dynamic' => true } }
+
+    it_behaves_like 'any attribute set'
+  end
+
+  context 'an overridden environment attribute' do
+    let(:environment) { { 'IGNORE_RESOLVCONF' => 'yes', 'PANTS' => 'no' } }
 
     it_behaves_like 'any attribute set'
   end