+++ CHANGELOG.md
@@ -8,7 +8,8 @@ Unreleased
 - Convert all the HWRPs to Chef custom resources
 - Patch the custom resources for Chef 13 compatibility
 - Support multiple config resources with different filename properties
-- Work around an failure condition with the :remove action on RHEL6
+- Work around a failure condition with the :remove action on RHEL6
+- Add an environment property and attribute for passing to the service resource
 
 v1.1.0 (2016-08-11)
 -------------------